+++ hl/src/tcp.rs
@@ -58,7 +58,7 @@ use w5500_ll::{
 /// ```
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct TcpReader<'w, W5500: Registers> {
+pub struct TcpReader<'w, W5500> {
     pub(crate) w5500: &'w mut W5500,
     pub(crate) sn: Sn,
     pub(crate) head_ptr: u16,
@@ -66,8 +66,8 @@ pub struct TcpReader<'w, W5500: Registers> {
     pub(crate) ptr: u16,
 }
 
-impl<'w, W5500: Registers> Seek<W5500::Error> for TcpReader<'w, W5500> {
-    fn seek(&mut self, pos: SeekFrom) -> Result<(), Error<W5500::Error>> {
+impl<'w, W5500> Seek for TcpReader<'w, W5500> {
+    fn seek<E>(&mut self, pos: SeekFrom) -> Result<(), Error<E>> {
         self.ptr = pos.new_ptr(self.ptr, self.head_ptr, self.tail_ptr)?;
         Ok(())
     }
@@ -122,6 +122,47 @@ impl<'a, W5500: Registers> Read<W5500::Error> for TcpReader<'a, W5500> {
     }
 }
 
+#[cfg(feature = "async")]
+impl<'w, W5500: w5500_ll::aio::Registers + 'w> crate::io::AsyncRead<W5500::Error>
+    for TcpReader<'w, W5500>
+{
+    type ReadFuture<'a> = impl core::future::Future<Output = Result<u16, W5500::Error>> + 'a
+        where Self: 'a;
+
+    fn read<'a>(&'a mut self, buf: &'a mut [u8]) -> Self::ReadFuture<'a> {
+        async move {
+            let read_size: u16 = min(self.remain(), buf.len().try_into().unwrap_or(u16::MAX));
+            if read_size != 0 {
+                self.w5500
+                    .sn_rx_buf(self.sn, self.ptr, &mut buf[..usize::from(read_size)])
+                    .await?;
+                self.ptr = self.ptr.wrapping_add(read_size);
+
+                Ok(read_size)
+            } else {
+                Ok(0)
+            }
+        }
+    }
+
+    type ReadExactFuture<'a> = impl core::future::Future<Output = Result<(), Error<W5500::Error>>> + 'a
+        where Self: 'a;
+
+    fn read_exact<'a>(&'a mut self, buf: &'a mut [u8]) -> Self::ReadExactFuture<'a> {
+        async move {
+            let buf_len: u16 = buf.len().try_into().unwrap_or(u16::MAX);
+            let write_size: u16 = min(self.remain(), buf_len);
+            if write_size != buf_len {
+                Err(Error::OutOfMemory)
+            } else {
+                self.w5500.set_sn_tx_buf(self.sn, self.ptr, buf).await?;
+                self.ptr = self.ptr.wrapping_add(write_size);
+                Ok(())
+            }
+        }
+    }
+}
+
 /// Streaming writer for a UDP socket buffer.
 ///
 /// This implements the [`Seek`] traits.
@@ -158,7 +199,7 @@ impl<'a, W5500: Registers> Read<W5500::Error> for TcpReader<'a, W5500> {
 /// ```
 #[derive(Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct TcpWriter<'w, W5500: Registers> {
+pub struct TcpWriter<'w, W5500> {
     pub(crate) w5500: &'w mut W5500,
     pub(crate) sn: Sn,
     pub(crate) head_ptr: u16,
@@ -166,8 +207,8 @@ pub struct TcpWriter<'w, W5500: Registers> {
     pub(crate) ptr: u16,
 }
 
-impl<'w, W5500: Registers> Seek<W5500::Error> for TcpWriter<'w, W5500> {
-    fn seek(&mut self, pos: SeekFrom) -> Result<(), Error<W5500::Error>> {
+impl<'w, W5500> Seek for TcpWriter<'w, W5500> {
+    fn seek<E>(&mut self, pos: SeekFrom) -> Result<(), Error<E>> {
         self.ptr = pos.new_ptr(self.ptr, self.head_ptr, self.tail_ptr)?;
         Ok(())
     }
@@ -219,6 +260,47 @@ impl<'w, W5500: Registers> Write<W5500::Error> for TcpWriter<'w, W5500> {
         self.w5500.set_sn_tx_wr(self.sn, self.ptr)?;
         self.w5500.set_sn_cr(self.sn, SocketCommand::Send)?;
         Ok(())
+    }
+}
+
+#[cfg(feature = "async")]
+impl<'w, W5500: w5500_ll::aio::Registers + 'w> crate::io::AsyncWrite<W5500::Error>
+    for TcpWriter<'w, W5500>
+{
+    type WriteFuture<'a> = impl core::future::Future<Output = Result<u16, W5500::Error>> + 'a
+        where Self: 'a;
+
+    fn write<'a>(&'a mut self, buf: &'a [u8]) -> Self::WriteFuture<'a> {
+        async move {
+            let write_size: u16 = min(self.remain(), buf.len().try_into().unwrap_or(u16::MAX));
+            if write_size != 0 {
+                self.w5500
+                    .set_sn_tx_buf(self.sn, self.ptr, &buf[..usize::from(write_size)])
+                    .await?;
+                self.ptr = self.ptr.wrapping_add(write_size);
+
+                Ok(write_size)
+            } else {
+                Ok(0)
+            }
+        }
+    }
+
+    type WriteAllFuture<'a> = impl core::future::Future<Output = Result<(), Error<W5500::Error>>> + 'a
+        where Self: 'a;
+
+    fn write_all<'a>(&'a mut self, buf: &'a [u8]) -> Self::WriteAllFuture<'a> {
+        async move {
+            let buf_len: u16 = buf.len().try_into().unwrap_or(u16::MAX);
+            let write_size: u16 = min(self.remain(), buf_len);
+            if write_size != buf_len {
+                Err(Error::OutOfMemory)
+            } else {
+                self.w5500.set_sn_tx_buf(self.sn, self.ptr, buf).await?;
+                self.ptr = self.ptr.wrapping_add(write_size);
+                Ok(())
+            }
+        }
     }
 }